Modularizing the bot a little bit
added creds.py for keys and such 
changed structure of main.py for modularity 
added contentscheduler module
